Moved and organized files of tauri translation system
